Update Constants.java
I edited the constants for neo base robot. We can set max speed and acceleration now along with more crucial things. We will have to measure the wheel diameter and find Angle and Drive Motors pid values. After that we have to set drive gear ratio(I didn't put it in the code yet) and we will be all done. Also there's no chosen module for swerve so we have to do the calculations ourselves. None of the code was edited except the deletion of the talon swerve libary. All new code is commented out so it doesn't affect the structure of the file.